+++ mongosql/crud/__init__.py
@@ -1,2 +1,2 @@
 from .crudhelper import CrudHelper, StrictCrudHelper
-from .crudview import CrudViewMixin, CRUD_METHOD, saves_relations
+from .crudview import CrudViewMixin, CRUD_METHOD, saves_relations, ABSENT